refactor(tune): extract CPU tape + Adam optimizer into reusable module (#445)
- New crates/tune/src/lora/train_core.rs: shared public module containing
  Dims, GdnDims, MixerKind, LayerW, Head, LoraParams/Grads, SeqCtx, FullFwd,
  forward_full, eval_chain_nll, nll_and_grads, shifted, rand_fill, and
  apply_adam_updates. All public functions return Result<T> (no unwrap).
- mod.rs: declare train_core before train under the train-backward feature gate,
  with #[doc(hidden)] pub mod so the binary crate can import it.
- train.rs: replace ~500 lines of private type/function definitions with imports
  from train_core; inline Adam loop replaced by apply_adam_updates call.
- train_grad_full.rs: replace duplicate local definitions and Adam loop with the
  same train_core imports; all call sites updated to propagate Result with ?.

Zero behavior change: cargo test 17/17 pass (same as baseline).
